Move o.s.c.util to o.s.c.commons.util
